requirements.txt
# Task-2: Context-Aware Chatbot (LangChain/RAG + Streamlit)
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
chromadb>=0.4.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
pypdf>=3.0.0
tiktoken>=0.5.0
